src/main/webapp/app/entities/appreciations/delete/appreciations-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { AppreciationsService } from '../service/appreciations.service';

import { AppreciationsDeleteDialogComponent } from './appreciations-delete-dialog.component';

describe('Component Tests', () => {
  describe('Appreciations Management Delete Component', () => {
    let comp: AppreciationsDeleteDialogComponent;
    let fixture: ComponentFixture<AppreciationsDeleteDialogComponent>;
    let service: AppreciationsService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [AppreciationsDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(AppreciationsDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(AppreciationsDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(AppreciationsService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete('ABC');
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith('ABC');
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
